app/server/python/requirements.txt
grpcio==1.42.0
google-api-python-client==2.30.0
grpcio-reflection==1.42.0
aioprometheus==21.9.1
aiohttp==3.8.1
pympler==1.0.1
